Rename Account interface and AbstractAccount class
