+++ src/main/java/com/codecoolbank/javase/controller/AccountController.java
@@ -1,0 +1,8 @@
+package com.codecoolbank.javase.model;
+
+public interface Account {
+    void deposit(long amount);
+    void withdraw(long amount);
+    long getBalance();
+    int getAccountId();
+}
